+++ src/components/calculator/components/CombinedMetrics/index.ts
@@ -2,4 +2,4 @@ export { CombinedMetrics } from './CombinedMetrics';
 export { MetricCard } from './MetricCard';
 export { SummaryItem } from './SummaryItem';
 export { TaxEquitySection } from './TaxEquitySection';
-export * from './utils';
+export { YearSelector } from './YearSelector';